+++ src/main/java/com/uqpay/sdk/banking/model/VirtualAccountApplicationResultStatus.java
@@ -1,0 +1,9 @@
+package com.uqpay.sdk.banking.model;
+
+public enum VirtualAccountApplicationResultStatus {
+    SUBMITTED,
+    COMPLETED,
+    FAILED,
+    SKIPPED,
+    CLOSED
+}