+++ src/main/java/com/uqpay/sdk/banking/model/VirtualAccountApplicationStatus.java
@@ -1,0 +1,9 @@
+package com.uqpay.sdk.banking.model;
+
+public enum VirtualAccountApplicationStatus {
+    SUBMITTED,
+    PARTIALLY_COMPLETED,
+    COMPLETED,
+    FAILED,
+    CLOSED
+}